webapp/requirements.txt
# Web Application Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
jinja2==3.1.2
python-multipart==0.0.6
pydantic==2.5.0

# Google Cloud
google-cloud-firestore==2.13.1
google-cloud-storage==2.13.0
google-cloud-run==0.10.1

# Parent package (includes all ML dependencies)
-e .
